+++ content/graphs/floydWarshall.h
@@ -2,9 +2,9 @@
 
 //status: tested against output of dijkstras on random graphs
 
-vector<vector<ll>> floydWarshall(const vector<vector<pair<int, ll>>>& adj /*directed or undirected, weighted graph*/) {
+vector<vector<long long>> floydWarshall(const vector<vector<pair<int, long long>>>& adj /*directed or undirected, weighted graph*/) {
 	int n = adj.size();
-	vector<vector<ll>> len(n, vector<ll> (n, 1e18));
+	vector<vector<long long>> len(n, vector<long long> (n, 1e18));
 	for (int i = 0; i < n; i++) {
 		len[i][i] = 0;//remove this line if you want shortest cycle - len[i][i] will = length of shortest cycle including node i (only for directed graphs)
 		for (auto [neighbor, weight] : adj[i])
