+++ content/math/exp_mod.h
@@ -3,8 +3,8 @@
 //status: tested on random inputs, and used in misc. problems
 
 //returns a^pw % mod in O(log(pw))
-ll fastPow(ll a, ll pw, int mod) {
-	ll res = 1;
+long long fastPow(long long a, long long pw, int mod) {
+	long long res = 1;
 	a %= mod;
 	while (pw > 0) {
 		if (pw & 1) res = (res * a) % mod;
